# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Auction Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /periods:
    get:
      operationId: getSportPeriods
      tags: [results]
      summary: List shared sport periods
      parameters:
        - name: sport
          in: query
          schema:
            type: string
            default: NFL
      responses:
        "200":
          description: Shared period template
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SportPeriod"
  /period-snapshots:
    post:
      operationId: upsertTeamPeriodSnapshot
      tags: [results]
      summary: Write one cumulative realized or mark-to-market team snapshot for an NFL period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamPeriodSnapshotInput"
      responses:
        "200":
          description: Upserted snapshot and calculated gross return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamPeriodSnapshotRow"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /period-snapshots/week-zero:
    post:
      operationId: initializeWeekZeroPoints
      tags: [results]
      summary: Initialize the immutable zero-stat Week 0 baseline for a season
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitializeWeekZeroPointsInput"
      responses:
        "200":
          description: Week 0 realized and MTM snapshots initialized or already present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekZeroPointsInitialization"
        "400":
          description: Validation error or no auctioned teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — ADMIN_API_KEY required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Season not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payout-rules:
    get:
      operationId: getPayoutRules
      tags: [results]
      summary: List a season Calcutta's payout rules
      parameters:
        - name: season
          in: query
          required: true
          schema:
            type: number
        - name: calcuttaId
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Payout rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayoutRuleRow"
    put:
      operationId: replacePayoutRules
      tags: [results]
      summary: Replace all payout rules for a season's canonical Calcutta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutRulesUpdate"
      responses:
        "200":
          description: Saved payout rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayoutRuleRow"
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      description: Administrative bearer token supplied through ADMIN_API_KEY.
    JobRunnerBearerAuth:
      type: http
      scheme: bearer
      description: Restricted job trigger token supplied through JOB_RUNNER_SECRET.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SportPeriod:
      type: object
      properties:
        sequence:
          type: number
          multipleOf: 1
        label:
          type: string
        isPlayoff:
          type: boolean
      required: [sequence, label, isPlayoff]
    PayoutMetric:
      type: string
      enum:
        - win
        - tie
        - pt_diff
        - playoff_berth
        - div_round
        - conf_round
        - sb_berth
        - win_super_bowl
    PayoutRuleInput:
      type: object
      properties:
        metric:
          $ref: "#/components/schemas/PayoutMetric"
        dollarsPerUnit:
          type: number
        playoffMultiplier:
          type: number
          minimum: 0
          default: 2
      required: [metric, dollarsPerUnit]
    PayoutRuleRow:
      type: object
      properties:
        metric:
          $ref: "#/components/schemas/PayoutMetric"
        dollarsPerUnit:
          type: number
        playoffMultiplier:
          type: number
          minimum: 0
      required: [metric, dollarsPerUnit, playoffMultiplier]
    PayoutRulesUpdate:
      type: object
      properties:
        seasonYear:
          type: number
        calcuttaId:
          type: number
        rules:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PayoutRuleInput"
      required: [seasonYear, rules]
    TeamPeriodSnapshotInput:
      type: object
      properties:
        teamId:
          type: number
        seasonYear:
          type: number
        calcuttaId:
          type: number
        periodSequence:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 22
        basis:
          type: string
          enum: [realized, mtm]
        wins:
          type: number
          minimum: 0
        losses:
          type: number
          minimum: 0
        ties:
          type: number
          minimum: 0
        ordinaryWins:
          type: number
          minimum: 0
        marqueeWins:
          type: number
          minimum: 0
        ordinaryTies:
          type: number
          minimum: 0
        marqueeTies:
          type: number
          minimum: 0
        ordinaryPtDiff:
          type: number
        marqueePtDiff:
          type: number
        ptDiff:
          type: number
        playoffBerth:
          type: number
          minimum: 0
          maximum: 1
        divRound:
          type: number
          minimum: 0
          maximum: 1
        confRound:
          type: number
          minimum: 0
          maximum: 1
        sbBerth:
          type: number
          minimum: 0
          maximum: 1
        winSuperBowl:
          type: number
          minimum: 0
          maximum: 1
        playoffStatus:
          type: string
          enum: [unknown, alive, clinched, eliminated]
      required: [teamId, seasonYear, periodSequence, basis]
    TeamPeriodSnapshotRow:
      type: object
      properties:
        teamId:
          type: number
        seasonYear:
          type: number
        periodSequence:
          type: number
          multipleOf: 1
          minimum: 0
          maximum: 22
        periodLabel:
          type: string
        isPlayoff:
          type: boolean
        basis:
          type: string
          enum: [realized, mtm]
        wins:
          type: number
        losses:
          type: number
        ties:
          type: number
        ordinaryWins:
          type: number
        marqueeWins:
          type: number
        ordinaryTies:
          type: number
        marqueeTies:
          type: number
        ordinaryPtDiff:
          type: number
        marqueePtDiff:
          type: number
        ptDiff:
          type: number
        playoffBerth:
          type: number
        divRound:
          type: number
        confRound:
          type: number
        sbBerth:
          type: number
        winSuperBowl:
          type: number
        playoffStatus:
          type: string
          enum: [unknown, alive, clinched, eliminated]
        grossReturn:
          type: number
      required:
        - teamId
        - seasonYear
        - periodSequence
        - periodLabel
        - isPlayoff
        - basis
        - wins
        - losses
        - ties
        - ordinaryWins
        - marqueeWins
        - ordinaryTies
        - marqueeTies
        - ordinaryPtDiff
        - marqueePtDiff
        - ptDiff
        - playoffBerth
        - divRound
        - confRound
        - sbBerth
        - winSuperBowl
        - playoffStatus
        - grossReturn
    InitializeWeekZeroPointsInput:
      type: object
      properties:
        seasonYear:
          type: number
          description: NFL season to initialize
        calcuttaId:
          type: number
      required: [seasonYear]
    WeekZeroPointsInitialization:
      type: object
      properties:
        seasonYear:
          type: number
        periodSequence:
          type: number
        periodLabel:
          type: string
        teamCount:
          type: number
        realizedSnapshotsWritten:
          type: number
        mtmSnapshotsWritten:
          type: number
        snapshotsWritten:
          type: number
        alreadyInitialized:
          type: boolean
      required:
        - seasonYear
        - periodSequence
        - periodLabel
        - teamCount
        - realizedSnapshotsWritten
        - mtmSnapshotsWritten
        - snapshotsWritten
        - alreadyInitialized
...