# lib/api-zod/src/generated/types/initializeWeekZeroPointsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NFL Auction Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface InitializeWeekZeroPointsInput {
  /** NFL season to initialize */
  seasonYear: number;
  calcuttaId?: number;
}