# lib/api-zod/src/generated/types/payoutRulesUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NFL Auction Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { PayoutRuleInput } from './payoutRuleInput';

export interface PayoutRulesUpdate {
  seasonYear: number;
  calcuttaId?: number;
  /** @minItems 1 */
  rules: PayoutRuleInput[];
}